sound effects! map resources! backgrounds! fixed DMA!

+++ common/src/md/psg.h
@@ -36,8 +36,8 @@ Michael Moffitt 2018 */
 
 // Megadrive PSG functions
 static inline void psg_vol(uint8_t chan, uint8_t vol);
-static inline void psg_pitch(uint8_t chan, uint32_t pitch);
-static inline void psg_tone(uint8_t chan, uint8_t vol, uint32_t pitch);
+static inline void psg_pitch(uint8_t chan, uint16_t pitch);
+static inline void psg_tone(uint8_t chan, uint8_t vol, uint16_t pitch);
 static inline void psg_note(uint8_t chan, uint8_t note, uint8_t octave);
 
 
@@ -46,13 +46,13 @@ static inline void psg_vol(uint8_t chan, uint8_t vol)
 	PSG_PORT = 0x90 | ((chan & 0x03) << 5) | (vol & 0x0F);
 }
 
-static inline void psg_pitch(uint8_t chan, uint32_t pitch)
+static inline void psg_pitch(uint8_t chan, uint16_t pitch)
 {
 	PSG_PORT = 0x80 | ((chan & 0x03) << 5) | (pitch & 0x0F);
 	PSG_PORT = (pitch >> 4) & 0x3F;
 }
 
-static inline void psg_tone(uint8_t chan, uint8_t vol, uint32_t pitch)
+static inline void psg_tone(uint8_t chan, uint8_t vol, uint16_t pitch)
 {	
 	psg_pitch(chan,pitch);
 	psg_vol(chan,vol);
